extract fixture
